Remove exchange_result.hpp
